+++ MogaStyle/app/src/main/java/com/example/mogastyle/Activities/Consult/ConsultFragment.java
@@ -1,4 +1,4 @@
-package com.example.mogastyle.Activities;
+package com.example.mogastyle.Activities.Consult;
 
 import android.os.Bundle;
 
@@ -12,10 +12,10 @@ import com.example.mogastyle.R;
 
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link MyPageFragment#newInstance} factory method to
+ * Use the {@link ConsultFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class MyPageFragment extends Fragment {
+public class ConsultFragment extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -26,7 +26,7 @@ public class MyPageFragment extends Fragment {
     private String mParam1;
     private String mParam2;
 
-    public MyPageFragment() {
+    public ConsultFragment() {
         // Required empty public constructor
     }
 
@@ -36,11 +36,11 @@ public class MyPageFragment extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment MyPageFragment.
+     * @return A new instance of fragment ConsultFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static MyPageFragment newInstance(String param1, String param2) {
-        MyPageFragment fragment = new MyPageFragment();
+    public static ConsultFragment newInstance(String param1, String param2) {
+        ConsultFragment fragment = new ConsultFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -61,6 +61,6 @@ public class MyPageFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_my_page, container, false);
+        return inflater.inflate(R.layout.fragment_consult, container, false);
     }
 }
