+++ MogaStyle/app/src/main/java/com/example/mogastyle/Activities/LoginActivity.java
@@ -7,15 +7,39 @@ package com.example.mogastyle.Activities;
 
 import androidx.appcompat.app.AppCompatActivity;
 
+import android.content.Intent;
 import android.os.Bundle;
+import android.view.View;
+import android.widget.Button;
 
 import com.example.mogastyle.R;
 
 public class LoginActivity extends AppCompatActivity {
 
+    Button btn_login_goMain;
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_login);
+
+        btn_login_goMain = findViewById(R.id.btn_login_goMain);
+        btn_login_goMain.setOnClickListener(onClickListener);
+
     }
+
+    View.OnClickListener onClickListener = new View.OnClickListener() {
+        @Override
+        public void onClick(View v) {
+            switch(v.getId()){
+                case R.id.btn_login_goMain:
+                    Intent intent = new Intent(LoginActivity.this , MainActivity.class);
+                    startActivity(intent);
+                    break;
+
+
+            }
+        }
+    };
+
 }
