+++ MogaStyle/app/src/main/res/layout/fragment_hair_shop.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Activities.HairShopFragment">
+    tools:context=".Activities.HairShop.HairShopFragment">
 
 
     <TextView
