+++ MogaStyle/app/src/main/res/layout/fragment_my_page.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Activities.MyPageFragment">
+    tools:context=".Activities.MyPage.MyPageFragment">
 
 
     <TextView
